spawning of tower is no longer managed in the player script itself.
